Refactor JSON parsing logic into ChatViewModel
- Extracted JSON parsing of news conversations out of `ChatDetailFragment` and into `ChatViewModel`.
- Injected `DispatcherProvider` into `ChatViewModel` (adhering to architecture guidelines) to run the parsing safely on the IO thread via `withContext`.
- Updated `ChatViewModelTest` with `TestDispatcherProvider` to fix test dependencies.
- Verified that `onDestroyView` already properly detaches the `TextWatcher` inside an `isInitialized` guard, so no changes were needed there.
