+++ app/src/main/java/org/ole/planet/myplanet/ui/chat/ChatDetailFragment.kt
@@ -47,7 +47,6 @@ import org.ole.planet.myplanet.services.SharedPrefManager
 import org.ole.planet.myplanet.services.sync.ServerUrlMapper
 import org.ole.planet.myplanet.ui.dashboard.DashboardActivity
 import org.ole.planet.myplanet.utils.DialogUtils
-import org.ole.planet.myplanet.utils.DispatcherProvider
 import org.ole.planet.myplanet.utils.JsonUtils
 import retrofit2.Response
 
@@ -71,8 +70,6 @@ class ChatDetailFragment : Fragment() {
     @Inject
     lateinit var sharedPrefManager: SharedPrefManager
     lateinit var customProgressDialog: DialogUtils.CustomProgressDialog
-    @Inject
-    lateinit var dispatcherProvider: DispatcherProvider
     @Inject
     lateinit var chatRepository: ChatRepository
     @Inject
@@ -207,17 +204,7 @@ class ChatDetailFragment : Fragment() {
             customProgressDialog.setText(getString(R.string.please_wait))
             customProgressDialog.show()
             try {
-                val messages = withContext(dispatcherProvider.io) {
-                    val conversations = JsonUtils.gson.fromJson(newsConversations, Array<RealmConversation>::class.java).toList()
-                    val list = mutableListOf<ChatMessage>()
-                    val limit = 20
-                    val limitedConversations = if (conversations.size > limit) conversations.takeLast(limit) else conversations
-                    for (conversation in limitedConversations) {
-                        conversation.query?.let { list.add(ChatMessage(it, ChatMessage.QUERY)) }
-                        conversation.response?.let { list.add(ChatMessage(it, ChatMessage.RESPONSE, ChatMessage.RESPONSE_SOURCE_SHARED_VIEW_MODEL)) }
-                    }
-                    list
-                }
+                val messages = sharedViewModel.parseNewsConversations(newsConversations)
                 mAdapter.submitList(messages) {
                     binding.recyclerGchat.post {
                         binding.recyclerGchat.scrollToPosition(mAdapter.itemCount - 1)
